Upgrade libft and add new functions from  projects ft_printf and push_swap.

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: gly <marvin@42.fr>                         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/11/06 13:36:34 by gly               #+#    #+#             */
-/*   Updated: 2018/11/12 17:59:12 by gly              ###   ########.fr       */
+/*   Updated: 2018/11/20 11:03:00 by gly              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include "libft.h"
+#include <string.h>
 
 size_t	ft_strlen(char const *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
